+++ core/samples/reflection/traits/constructible/main.cpp
@@ -7,10 +7,17 @@ struct Scale
 {
     CUBOS_REFLECT;
     float value = 1.0F;
+
+    Scale() = default;
+    Scale(float value)
+        : value(value)
+    {
+    }
 };
 /// [Scale declaration]
 
 /// [Scale definition]
+#include <cubos/core/reflection/external/primitives.hpp>
 #include <cubos/core/reflection/traits/constructible.hpp>
 #include <cubos/core/reflection/type.hpp>
 
@@ -19,15 +26,16 @@ using cubos::core::reflection::Type;
 
 CUBOS_REFLECT_IMPL(Scale)
 {
-    return Type::create("Scale").with(ConstructibleTrait::typed<Scale>().withDefaultConstructor().build());
+    return Type::create("Scale").with(
+        ConstructibleTrait::typed<Scale>().withDefaultConstructor().withCustomConstructor<float>({"value"}).build());
 }
 /// [Scale definition]
 
-/// [Accessing the trait]
 int main()
 {
     using cubos::core::reflection::reflect;
 
+    /// [Accessing the trait]
     const auto& scaleType = reflect<Scale>();
     CUBOS_ASSERT(scaleType.has<ConstructibleTrait>());
     const auto& constructible = scaleType.get<ConstructibleTrait>();
@@ -43,6 +51,20 @@ int main()
     /// [Destroying the instance]
     // Destroy the instance and deallocate its memory.
     constructible.destruct(instance);
+    /// [Destroying the instance]
+
+    /// [Creating a custom instance]
+    CUBOS_ASSERT(constructible.customConstructorCount() == 1);
+    CUBOS_ASSERT(constructible.customConstructor(0).argCount() == 1);
+    CUBOS_ASSERT(constructible.customConstructor(0).argName(0) == "value");
+    CUBOS_ASSERT(constructible.customConstructor(0).argType(0).is<float>());
+
+    float value = 2.0F;
+    void* args[] = {&value};
+    constructible.customConstruct(0, instance, args);
+    CUBOS_ASSERT(static_cast<Scale*>(instance)->value == 2.0F);
+    /// [Creating a custom instance]
+
+    constructible.destruct(instance);
     operator delete(instance);
 }
-/// [Destroying the instance]