+++ core/src/reflection/traits/constructible.cpp
@@ -9,6 +9,45 @@ CUBOS_REFLECT_IMPL(ConstructibleTrait)
     return Type::create("cubos::core::reflection::ConstructibleTrait");
 }
 
+ConstructibleTrait::CustomConstructor::CustomConstructor(
+    std::vector<const Type*> argTypes, std::vector<std::string> argNames,
+    memory::Function<void(void* instance, void** args) const> functor)
+    : mArgTypes(memory::move(argTypes))
+    , mArgNames(memory::move(argNames))
+    , mFunctor(memory::move(functor))
+{
+    CUBOS_ASSERT(mFunctor != nullptr, "Functor must be non-null");
+    CUBOS_ASSERT(mArgTypes.size() == mArgNames.size(), "Argument types and names must have the same size");
+    for (const Type* argType : mArgTypes)
+    {
+        CUBOS_ASSERT(argType != nullptr, "Argument type must not be null");
+    }
+}
+
+std::size_t ConstructibleTrait::CustomConstructor::argCount() const
+{
+    return mArgTypes.size();
+}
+
+const std::string& ConstructibleTrait::CustomConstructor::argName(std::size_t index) const
+{
+    CUBOS_ASSERT(index < mArgNames.size(), "Index out of bounds");
+    return mArgNames[index];
+}
+
+const Type& ConstructibleTrait::CustomConstructor::argType(std::size_t index) const
+{
+    CUBOS_ASSERT(index < mArgTypes.size(), "Index out of bounds");
+    return *mArgTypes[index];
+}
+
+void ConstructibleTrait::CustomConstructor::call(void* instance, void** args) const
+{
+    CUBOS_ASSERT(instance != nullptr, "Instance must not be null");
+    CUBOS_ASSERT(args != nullptr, "Args must not be null");
+    mFunctor(instance, args);
+}
+
 ConstructibleTrait::ConstructibleTrait(std::size_t size, std::size_t alignment, Destructor destructor)
     : mSize(size)
     , mAlignment(alignment)
@@ -40,6 +79,17 @@ ConstructibleTrait&& ConstructibleTrait::withMoveConstructor(MoveConstructor mov
     return memory::move(*this);
 }
 
+ConstructibleTrait&& ConstructibleTrait::withCustomConstructor(CustomConstructor customConstructor) &&
+{
+    this->addCustomConstructor(memory::move(customConstructor));
+    return memory::move(*this);
+}
+
+void ConstructibleTrait::addCustomConstructor(CustomConstructor customConstructor)
+{
+    mCustomConstructors.emplace_back(memory::move(customConstructor));
+}
+
 std::size_t ConstructibleTrait::size() const
 {
     return mSize;
@@ -65,6 +115,11 @@ bool ConstructibleTrait::hasMoveConstruct() const
     return mMoveConstructor != nullptr;
 }
 
+std::size_t ConstructibleTrait::customConstructorCount() const
+{
+    return mCustomConstructors.size();
+}
+
 void ConstructibleTrait::destruct(void* instance) const
 {
     mDestructor(instance);
@@ -88,6 +143,11 @@ void ConstructibleTrait::moveConstruct(void* instance, void* other) const
     mMoveConstructor(instance, other);
 }
 
+void ConstructibleTrait::customConstruct(std::size_t index, void* instance, void** args) const
+{
+    customConstructor(index).call(instance, args);
+}
+
 ConstructibleTrait::Destructor ConstructibleTrait::destructor() const
 {
     return mDestructor;
@@ -106,4 +166,10 @@ ConstructibleTrait::CopyConstructor ConstructibleTrait::copyConstructor() const
 ConstructibleTrait::MoveConstructor ConstructibleTrait::moveConstructor() const
 {
     return mMoveConstructor;
+}
+
+const ConstructibleTrait::CustomConstructor& ConstructibleTrait::customConstructor(std::size_t index) const
+{
+    CUBOS_ASSERT(index < mCustomConstructors.size(), "Index out of bounds");
+    return mCustomConstructors[index];
 }